Revert "Merging development into personal development branch (#121)"
This reverts commit fc5a529adeeb10e492ad9a21bec24885514aadd4.
