--- dev_helpers/1point.csv
@@ -1,1 +1,0 @@
-0,1,1,1
