--- dev_helpers/2points.csv
@@ -1,2 +1,0 @@
-0.01,0.01,0.01
-1,1,1
