--- dev_helpers/newWorldTestInput.csv
@@ -1,4 +1,0 @@
-"index", "x", "y", "z"
-1,1,1,1
-2,2,2,2
-3,3,3,3
