--- dev_helpers/ripple2-100.csv
@@ -1,100 +1,0 @@
-0,6.134542808632364,5.0403509580003405,1.567590421821693
-1,9.717091512674273,5.110602930004269,0.7801498032431362
-2,1.2473047671123572,5.605643510037619,0.9347096598988046
-3,6.219765126329638,1.092133042475656,0.168188828278836
-4,0.6904272893181873,8.247467181876567,2.337668799133397
-5,6.90634225786736,5.1965301045171355,3.122128724020373
-6,8.596655883080915,1.2809999584238274,3.213189887933076
-7,7.457512427388585,3.7370969176683766,2.4856720052860575
-8,1.8713311329028126,4.234257932329514,3.3820276561431166
-9,9.669547967618668,8.021034341481686,1.45717299454548
-10,8.895976452662465,1.582836107087029,3.7165401352928487
-11,3.4430745642410923,7.925888145681302,3.119131492677676
-12,4.924263032513425,2.065723124424077,1.7362035963551803
-13,1.1283691111052807,7.2722612227134835,0.47324581217266914
-14,7.51039786404176,1.8138698036943512,1.1124299227796512
-15,2.224632625495526,3.9468228199239874,3.6151130909038436
-16,4.096248006427218,6.844727174844689,1.6509007981044599
-17,1.4558475429726747,7.727350233110608,1.399698751802347
-18,0.5635825103191028,8.778800399154964,3.395570525255
-19,1.3686781492241695,4.749842802671137,2.63948053668
-20,9.958875390241289,7.021702195871111,0.6080811656873759
-21,9.05572400225468,3.7093074679304916,3.608586998642246
-22,4.504669294082247,8.581895583522197,3.7227760418611378
-23,0.12641232892038445,7.485933073720058,0.6701272167830361
-24,7.07217514593483,6.813160518152727,3.5601103126262847
-25,1.4264132691996734,0.7032784823026672,8.46347962407333
-26,0.8107492707102781,9.835211916414469,3.4853186962597347
-27,0.643027772155017,8.78755499932215,3.4184048731568146
-28,6.591210539274824,2.2066850419113146,0.06646872652935291
-29,8.180056198685715,1.1700177905279951,2.3079047643670147
-30,7.607201749839193,1.3859003162455397,1.1247364601642316
-31,0.7168923985191544,0.6650391520153343,9.05842121563218
-32,3.651625783329976,4.914843942827036,0.36410703190646265
-33,8.833619580570137,2.3716794660670546,3.813760577173478
-34,6.697516760501809,9.709982886071394,0.06377678441766206
-35,2.23226473216106,1.5122465406382568,7.15552312401214
-36,8.422197297587289,9.284106710149974,1.390624792202152
-37,8.8754865837627,0.7310015768150357,3.559027404960976
-38,1.2402956091766015,4.07247262832493,4.249312605920515
-39,6.332550916169835,9.303592373708424,0.2675020291469
-40,4.896807461286693,6.361567137138603,1.7668055464539312
-41,2.0567262577919534,5.938434837497598,0.1945593249612704
-42,8.815647106845967,2.064641269177745,3.7352292827185227
-43,2.2284532393120458,2.547837830255469,6.04796924719539
-44,1.4756253205910308,5.1305157069612095,1.7394379460796316
-45,6.375910220690494,8.075113124156845,2.5863884374599437
-46,9.245760530303553,3.8768780280579787,3.1972207840791986
-47,4.91890969922379,8.873464025324843,2.935749973242724
-48,9.633881118510475,1.8455853574832692,3.5761870772738997
-49,0.6285215719990733,5.03653748470271,2.330468873629621
-50,2.561804244048407,3.725852661099286,3.6363130359173783
-51,8.875403266804726,3.5143478926025837,3.8441786474868005
-52,7.599726033255476,4.8729846524162435,3.7083113468733155
-53,2.4254713069601186,1.7341034862630245,6.7299606625018145
-54,4.560675777235053,9.881441023828296,1.0052871025274153
-55,7.908568562733617,0.3583691320003046,1.5166383292207186
-56,3.3919590583566195,4.035214805002246,1.8864658865781514
-57,3.938363937208872,7.85380962892148,3.3776986967051563
-58,2.0892195248576573,3.2400590976017876,5.129039450987146
-59,4.336992710569022,4.365792875899208,0.3279589845750245
-60,8.093356684585315,6.827497680330161,1.7808093925814221
-61,6.709351526453721,6.4248212273608605,3.8843391884785445
-62,4.912267464149921,7.969286094116368,3.8951245478813297
-63,0.2691800723570803,0.3959346890770732,9.531945869499813
-64,3.480189907262514,6.928092222117823,1.1668735980530298
-65,2.809413832155001,8.59423896276458,3.722792923349765
-66,0.274657250482252,5.5148193856703545,1.3551406600577018
-67,5.704391356155164,4.329034021459568,0.23114364739122664
-68,8.560743015703196,3.2415217556148246,3.8189933189209406
-69,7.930919845121273,8.548110629215083,0.0035690206028324784
-70,1.329640039890383,4.852757683108727,2.4324598010114573
-71,6.359618239887847,2.4131382414806346,0.00991692861895021
-72,7.53758050547754,2.768352278250299,1.7711522472431698
-73,1.9195678789905735,5.202670155109315,1.307291311965701
-74,1.5700049594605725,8.393771983791847,2.9153703425850095
-75,8.074865531857146,4.280805842801355,3.808644475680888
-76,4.952135588711939,9.087174448396953,2.4304455180453073
-77,9.52965537843746,5.791551141804786,0.4321778203119418
-78,6.831940759801602,9.442428126846066,0.002647358265862591
-79,8.011833954820007,6.574928176721389,2.3898461113282967
-80,2.661853505519467,3.1680257813989243,4.518131376944978
-81,5.940984789935541,8.252278000367808,2.882772476436518
-82,9.580546585554382,4.2248766686429615,2.103522226082627
-83,2.7812642041764013,2.452488043512362,5.430390585628838
-84,5.074281137291305,9.893584024312641,0.4914445318832525
-85,6.5261534629210205,0.2546447703941779,0.03477950616019151
-86,5.39375170739146,6.670115907891912,2.994404296898016
-87,5.12785398604199,0.14204757258858725,2.20574406388811
-88,8.450010390269323,7.182500932498064,0.5465768601904825
-89,8.247844593038888,1.609972130426901,2.62416681182999
-90,1.4917707797296942,0.3070696581158039,8.531221761911995
-91,8.047877723691867,4.267608053194706,3.7851983658911568
-92,6.4811264004593525,2.7272954166829386,0.11813092125351488
-93,9.07530654736048,4.556186210263626,2.91450167227469
-94,4.977229670080869,8.130234731125807,3.851558713771916
-95,8.664365618098856,3.676956889060232,3.8924326606025006
-96,7.543338681075328,5.082102736399099,3.77353654417302
-97,8.868475001721821,2.1115450801962243,3.790906800414149
-98,9.964838927253512,7.171388045856122,0.7977936969211694
-99,7.268924183761504,3.408839525356094,1.7681088794188073
