+++ functions/package.json
@@ -5,7 +5,7 @@
   "main": "app.js",
   "scripts": {
     "start": "node ./bin/www",
-    "test": "nyc mocha --timeout 10000"
+    "test": "mocha --exit"
   },
   "repository": {
     "type": "git",
@@ -30,8 +30,7 @@
     "firebase-tools": "^3.17.4",
     "morgan": "~1.9.0",
     "selenium-webdriver": "^3.6.0",
-    "sinon": "^4.4.2",
-    "three-js": "^79.0.0"
+    "sinon": "^4.4.2"
   },
   "devDependencies": {
     "chai": "^4.1.2",
@@ -39,8 +38,6 @@
     "eslint-config-google": "^0.9.1",
     "firebase-server": "^0.12.0",
     "geckodriver": "^1.10.0",
-    "mocha": "^5.0.0",
-    "three": "^0.90.0",
-    "nyc": "^11.6.0"
+    "mocha": "^5.0.0"
   }
 }
