+++ public/javascripts/pointSelection.js
@@ -8,69 +8,25 @@
 
 var selectedPoints = [];  //array containing the indices of every currently
                           //selected point.
-var hiddenPoints = [];
-
+var pointSelectionRaycaster;
 var pointSelectionMouse = new THREE.Vector2();
-var selectionThreshold = 0.04; //the distance the mouse has to be from a point
+var selectionThreshold = 0.1; //the distance the mouse has to be from a point
 //in order for it to register as selectable
 var intersects;
 
-
-var selectionControllerL;
-var selectionControllerR;
-
-var pointSelectionRaycasterL;
-var pointSelectionRaycasterR;
-var raycasterLine;
-var isRaycasterLineActive;
 /**
  * Initializes the event listeners for point selection
  */
-
 function initializeSelectionControls()
 {
-  isRaycasterLineActive = false;
-  console.log("Initializing Selection Controls ... Point Selection Threshold: " + selectionThreshold);
+  console.log("Point Selection Threshold: " + selectionThreshold);
   pointSelectionRaycaster = new THREE.Raycaster();
   pointSelectionRaycaster.params.Points.threshold = selectionThreshold;
 
-  selectionControllerL = scene.getObjectByName("Oculus Touch (Left)");
-  selectionControllerR = scene.getObjectByName("Oculus Touch (Right)");
-
-  // We were originally going to allow selection with the left controller,
-  // But I think we'll probably limit it to the right control, and have tools
-  // on this one.
-  if (selectionControllerL){
-    console.log("Left VR Controller detected for point selection.");
-    console.log(selectionControllerL);
-    pointSelectionRaycasterL = new THREE.Raycaster();
-    pointSelectionRaycasterL.params.Points.threshold = selectionThreshold;
+  if (controller != null)
+  {
+    // TODO Attach raycaster to VRcontroller
   }
-
-  if (selectionControllerR){
-    console.log("Right VR Controller detected for point selection.");
-    console.log(selectionControllerR);
-    pointSelectionRaycasterR = new THREE.Raycaster();
-    pointSelectionRaycasterR.params.Points.threshold = selectionThreshold;
-    selectionControllerR.addEventListener('A touch began', function(event) {
-      isRaycasterLineActive = true;
-    });
-    selectionControllerR.addEventListener('A touch ended', function(event) {
-      isRaycasterLineActive = false;
-    });
-  }
-
-    /*
-    raycasterLineMaterial = new THREE.LineBasicMaterial({
-      color: 0xff0000
-    });
-    raycasterLineGeometry = new THREE.Geometry();
-    raycasterLineGeometry.vertices.push(controller.position);
-    raycasterLineGeometry.vertices.push(controller.position + (controller.rotation * 10));
-    raycasterLine = new THREE.Line(raycasterLineGeometry, raycasterLineMaterial);
-    scene.add(raycasterLine);
-    */
-
   // setup mouse raycaster here
 
   document.addEventListener( 'mousemove', onMouseMove, false );
@@ -82,89 +38,33 @@ function initializeSelectionControls()
  * frame.
  */
 var mousedOverPoint;
-var arrow;
-function pointSelectionUpdate() {
+function pointSelectionUpdate()
+{
   // calculate objects intersecting the ray
-
-  /*if (selectionControllerL){
-    pointSelectionRaycasterL.set(selectionControllerL.position, selectionControllerL.rotation);
-  }*/
-  if (selectionControllerR) {
-    var matrix = new THREE.Matrix4();
-    matrix.extractRotation( selectionControllerR.matrix );
-
-    var direction = new THREE.Vector3( 0, 0, 1 );
-    direction.applyMatrix4(matrix);
-    //matrix.multiplyVector3( direction );
-    direction.multiplyScalar(-1);
-    pointSelectionRaycasterR.set(selectionControllerR.position, direction);
-    intersects = pointSelectionRaycasterR.intersectObject(pointsSystem)
-
-    //console.log(selectionControllerR.getAxis(0));
-  }
-
-  // If no controllers are present, revert to mouse/camera selection.
-  if (selectionControllerL == null && selectionControllerR == null) {
-    pointSelectionRaycaster.setFromCamera(pointSelectionMouse, camera);
-    intersects = pointSelectionRaycaster.intersectObject(pointsSystem);
-  }
-
-
-  if (intersects != null) {
-    intersects = (intersects.length) > 0 ? intersects[0] : null;
-  }
-
-  // Reset point size when not moused over
+  pointSelectionRaycaster.setFromCamera(pointSelectionMouse, camera);
+  intersects = pointSelectionRaycaster.intersectObject( pointsSystem );
+  intersects = ( intersects.length ) > 0 ? intersects[ 0 ] : null;
   setPointScale(mousedOverPoint, plotPointSizeCoeff * Math.max(plotInitSizeX, plotInitSizeY, plotInitSizeZ));
   //pointsGeometry.boundingBox = null;
   if (intersects != null) {
     //console.log(intersects.point.x + " " + intersects.point.y + " " + intersects.point.z);
     //console.log(intersects);
-      if(pointsGeometry.getAttribute('isHidden').array[intersects.index] !== 1) {
-          setPointScale(intersects.index, plotPointSizeCoeff * Math.max(plotInitSizeX, plotInitSizeY, plotInitSizeZ) * 2);
+    setPointScale(intersects.index, plotPointSizeCoeff * Math.max(plotInitSizeX, plotInitSizeY, plotInitSizeZ) * 2);
 
-      }
-      mousedOverPoint = intersects.index;
-
-
-
+    mousedOverPoint = intersects.index;
+    //var curColor = getPointColor(intersects.index);
+    //pointsGeometry.getAttribute('customColor').array[intersects.index * 3] = 1;
+    //pointsGeometry.getAttribute('customColor').array[(intersects.index * 3) + 1] = 1;
+    //pointsGeometry.getAttribute('customColor').array[(intersects.index * 3) + 2] = 1;
+    //console.log(getPointColor(intersects.index));
+    //setPointColor(intersects.index, new THREE.Color(255,255,255));
+    //setPointColor(intersects.index, curColor.setHSL(curColor.getHSL.h, curColor.getHSL.s, curColor.getHSL.l * 1.5 ));
 
   }
   else {
 
     setPointScale(mousedOverPoint, plotPointSizeCoeff * Math.max(plotInitSizeX, plotInitSizeY, plotInitSizeZ));
   }
-
-  // Press 'A' and 'X' is select/deselect all points.
-
-  if (XisPressed && AisPressed){
-    XisPressed = false;
-    AisPressed = false;
-    if (selectedPoints.length > 0){
-      clearSelection();
-    }
-    else{
-      selectAll();
-    }
-  }
-
-
-
-  scene.remove ( raycasterLine );
-  if (pointSelectionRaycasterR && selectionControllerR && pointSelectionRaycasterR.ray.origin) {
-    var lineLength;
-    if (intersects){
-      lineLength = intersects.distance;
-    }
-    else {
-      lineLength = 1000000;
-    }
-    if (isRaycasterLineActive) {
-      raycasterLine = new THREE.ArrowHelper(pointSelectionRaycasterR.ray.direction, pointSelectionRaycasterR.ray.origin, lineLength, 0xff00ff, 0, 0);
-      scene.add(raycasterLine);
-    }
-  }
-
 }
 
 /**
@@ -174,10 +74,6 @@ function pointSelectionUpdate() {
  */
 function selectPoint(pointIndex)
 {
-  /*//make hidden points un-selectable
-  if(pointsGeometry.getAttribute('isHidden').array[pointIndex] === true){
-      return;
-  }*/
   pointsGeometry.getAttribute( 'isSelected' ).array[pointIndex] =
     !pointsGeometry.getAttribute( 'isSelected' ).array[pointIndex];
   if(pointsGeometry.getAttribute( 'isSelected' ).array[pointIndex] == false){
@@ -214,8 +110,6 @@ function clearSelection()
         pointsGeometry.getAttribute('position').array[(i*3)],
         pointsGeometry.getAttribute('position').array[(i*3)+1],
         pointsGeometry.getAttribute('position').array[(i*3)+2])));
-      setPointScale(i, pointsGeometry.getAttribute('size').array[i] =
-        plotPointSizeCoeff * Math.max(plotInitSizeX, plotInitSizeY, plotInitSizeZ));
     }
   }
   pointsGeometry.getAttribute( 'isSelected' ).array = selected;
@@ -267,16 +161,13 @@ function onClick( event ){
   event.preventDefault();
   if (intersects != null) {
     selectPoint(intersects.index);
-    //hidePoint(intersects.index);
   }
   else {
     clearSelection();
-    //unhideRecent();
   }
   if (selectedPoints.length > 0){
     console.log(getSelectedPointPositions());
   }
-  //console.log(hiddenPoints);
 }
 
 /**
@@ -347,8 +238,6 @@ function colorFromXYZcoords(vec3) {
  *
  * @return {Vector3[]} array of Vector3 objects containing positions
  */
-
-//must be fixed. Based on world position not value of points
 function getSelectedPointPositions() {
 
   var selectedPointPositions = [];
@@ -396,75 +285,10 @@ function getSelectedAxisValues(axis){
 }
 
 
-/////////////////////////////////////////////////////////////////////////
-///////////////////           POINT HIDING            ///////////////////
-/////////////////////////////////////////////////////////////////////////
 
 
-//if presses hide button away from a point, un-hides the most recently hidden point?
-
-function hidePoint(pointIndex){
-
-    pointsGeometry.getAttribute('isHidden').array[pointIndex] = true;
-    hiddenPoints.push(pointIndex);
-    //do the thing that hides it
-    //hiding by changing the colour to black is a very poor solution. Ideally color would include an alpha channel.
-    setPointColor(pointIndex, new THREE.Vector3(0,0,0));
 
 
-}
-
-function unhide(pointIndex){
-    hiddenPoints.splice(hiddenPoints.indexOf(pointIndex),1);
-    pointsGeometry.getAttribute('isHidden').array[pointIndex] = false;
-    //undo the thing that hides it
-    setPointColor(pointIndex, colorFromXYZcoords(new THREE.Vector3(
-        pointsGeometry.getAttribute('position').array[(pointIndex*3)],
-        pointsGeometry.getAttribute('position').array[(pointIndex*3)+1],
-        pointsGeometry.getAttribute('position').array[(pointIndex*3)+2])));
-
-}
-
-function unhideRecent(){
-    var recentIndex = (hiddenPoints[hiddenPoints.length -1]);
-    unhide(recentIndex);
-
-}
-
-function unhideAll(){
-    for(var i = 0; i < pointsGeometry.getAttribute('size').array.length; i++){
-        unhide(i);
-    }
-}
-
-
-function invertHidden(){
-  for( var i = 0; i < pointsGeometry.getAttribute('size').array.length; i++){
-    if(pointsGeometry.getAttribute('isHidden').array[i] === true){
-      unhide(i);
-    }
-    else {
-      hidePoint(i);
-    }
-  }
-    console.log(hiddenPoints);
-}
-
-
-function viewHidden(){
-    for( var i = 0; i < pointsGeometry.getAttribute('size').array.length; i++){
-        if(pointsGeometry.getAttribute('isHidden').array[i] === true || getPointColor(i) === new THREE.Color(0,0,0)){
-            setPointColor(i, colorFromXYZcoords(new THREE.Vector3(
-                pointsGeometry.getAttribute('position').array[(i*3)],
-                pointsGeometry.getAttribute('position').array[(i*3)+1],
-                pointsGeometry.getAttribute('position').array[(i*3)+2])));
-        }
-        else {
-            setPointColor(i, new THREE.Vector3(0,0,0));
-        }
-    }
-    console.log(hiddenPoints);
-}
